feat(articles): implement update article quantity functionality

+++ src/main/java/com/pragma/emazon/stock_microservice/domain/model/Article.java
@@ -8,14 +8,14 @@ public class Article {
     private Long id;
     private String name;
     private String description;
-    private Long quantity;
+    private Integer quantity;
     private BigDecimal price;
     private Brand brand;
     private List<Category> categories;
 
     public Article() {}
 
-    public Article(Long id, String name, String description, Long quantity, BigDecimal price, Brand brand, List<Category> categories) {
+    public Article(Long id, String name, String description, Integer quantity, BigDecimal price, Brand brand, List<Category> categories) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -49,11 +49,11 @@ public class Article {
         this.description = description;
     }
 
-    public Long getQuantity() {
+    public Integer getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(Long quantity) {
+    public void setQuantity(Integer quantity) {
         this.quantity = quantity;
     }
